rework reviews

+++ src/controllers/review_controller.go
@@ -9,23 +9,25 @@ import (
 )
 
 type ReviewControllerI interface {
-	CreateReview(reviewData models.CreateReviewRequest) (*uuid.UUID, *models.KTSError)
-	DeleteReview(id *uuid.UUID) *models.KTSError
+	CreateReview(reviewData models.CreateReviewRequest, userId *uuid.UUID) (string, *models.KTSError)
+	DeleteReview(id *uuid.UUID, userId *uuid.UUID) *models.KTSError
 }
 
 type ReviewController struct {
 	ReviewRepo repositories.ReviewRepositoryI
+	UserRepo   repositories.UserRepositoryI
 }
 
-func (rc ReviewController) CreateReview(reviewData models.CreateReviewRequest) (*uuid.UUID, *models.KTSError) {
+func (rc ReviewController) CreateReview(reviewData models.CreateReviewRequest, userId *uuid.UUID) (string, *models.KTSError) {
+	user, kts_err := rc.UserRepo.GetUserById(userId)
+	if kts_err != nil {
+		return "", kts_err
+	}
+
 	id := uuid.New()
 	movieId, err := uuid.Parse(reviewData.MovieID)
 	if err != nil {
-		return nil, kts_errors.KTS_BAD_REQUEST
-	}
-	userId, err := uuid.Parse(reviewData.UserID)
-	if err != nil {
-		return nil, kts_errors.KTS_BAD_REQUEST
+		return "", kts_errors.KTS_BAD_REQUEST
 	}
 
 	review := model.Reviews{
@@ -35,18 +37,25 @@ func (rc ReviewController) CreateReview(reviewData models.CreateReviewRequest) (
 		Datetime:  reviewData.Datetime,
 		IsSpoiler: &reviewData.IsSpoiler,
 		MovieID:   &movieId,
-		UserID:    &userId,
+		UserID:    userId,
 	}
 
 	kts_error := rc.ReviewRepo.CreateReview(review)
 	if kts_error != nil {
-		return nil, kts_error
+		return "", kts_error
 	}
 
-	return &id, nil
+	return *user.Username, nil
 }
 
-func (rc ReviewController) DeleteReview(id *uuid.UUID) *models.KTSError {
+func (rc ReviewController) DeleteReview(id *uuid.UUID, userId *uuid.UUID) *models.KTSError {
+	review, err := rc.ReviewRepo.GetReviewById(id)
+	if err != nil {
+		return err
+	}
+	if review.UserID != userId {
+		return kts_errors.KTS_FORBIDDEN
+	}
 	kts_error := rc.ReviewRepo.DeleteReview(id)
 	if kts_error != nil {
 		return kts_error
